feat: メソッドをモデル単位のファイルにまとめた
