+++ src/prisma/main.ts
@@ -1,2 +1,10 @@
 import "./client";
-export * from "./queries/main";
+export * from "./auth";
+
+export * from "./contests";
+
+export * from "./submissions";
+
+export * from "./users";
+
+export * from "./queue";
