--- src/prisma/queries/main.ts
@@ -1,11 +1,0 @@
-export * from "./auth/main";
-
-export * from "./contests/find";
-export * from "./contests/search";
-
-export * from "./submissions/create";
-export * from "./submissions/update";
-
-export * from "./users/find";
-
-export * from "./queue/create";
