+++ src/service/auth/main.ts
@@ -1,8 +1,7 @@
 import bcrypt from "bcrypt";
-// import * as fs from "fs";
 import jsonwebtoken from "jsonwebtoken";
-import { createNewAccount } from "../../prisma/queries/auth/main";
-import { findUserByName } from "../../prisma/queries/users/find";
+import { createNewAccount } from "../../prisma/auth";
+import { findUserByName } from "../../prisma/users";
 
 export async function registerUser(name: string, password: string) {
   const hashed_password = await bcrypt.hash(password, 10);
